+++ src/Entity/Interfaces/StationCloneAwareInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Entity\Interfaces;
+
+use App\Entity;
+
+interface StationCloneAwareInterface
+{
+    public function setStation(Entity\Station $station): void;
+}